Modified to calculate account balance using sumTransactions() of class Account.
In statementForAccount(), the process of calculating account balance has been modified so that the total is obtained by calling the account's sumTransactions() method, rather than by summing up the amounts in the loop through the list of transactions. This can be done because the Account class has been modified so that sumTransactions() returns the value of a member variable containing the account balance.