Commented out getFirstCustomer() function.
The function getFirstCustomer() appears to serve no useful purpose in the program. It wipes out the Bank object's list of Customers, then tries to do an access on what is now a null object, which will definitely throw an exception. Additionally, this function is not called anywhere in the program. For these reasons, the function has been commented out.